Publish release and main workflows to preview feed
Use matrix strategy for jobs that publish to the main NuGet feed,
both for prerelease and tagged release versions. Additionally run
the job to publish to the prerelease feed as well.

This allows targeting only the prerelease feed to get all packages,
including the final release versions. This simplifies configuration and
potential overlap of package for any consumers of the prerelease feed.